airflow/requirements.txt
kaggle==1.5.13
snowflake-connector-python==3.6.0
python-dotenv==1.0.0